requirements.txt
Flask-Migrate
flask_marshmallow
marshmallow-sqlalchemy
Flask-CORS
psycopg2-binary
langchain
python-dotenv
openai
flask-restx